Add support for expanded mode.
Changes the protocol for the screen buffer upate
is required since expanded mode can change during runtime,
so sending it with every screen update is the cleanest
way to add it.

Caveat: There is something funny with the data. When we switch
into expanded mode from normal mode the screen buffer looks very
different from when we pause and resume that same session.

Also cleans up the JS code quite a bit.