Support for N's Multi-Window
The emulator up until now used the decor view of the top-level
window to determine the screen dimensions for the emulated
TRS screen. With N's multi-window support, the decor view
is apparently covering the complete device's screen; even
though the app runs in a split window. This is strange. In order
to determine the proper window dimensions, the following trick
is used: first setContentView(R.layout.emulator_measure). This
is just a dummy view with one top-level view. On this view a
layout listener is set. When the layout finished computing,
the size of the dummy view is used for the TRS screen suze.
Only then the main layout R.layout.emulator is set. This seems
to be working reliably for all versions of Android. Sigh...