Enable direct buffer access to screen buffer.
 - A shared direct buffer for the screen is now used.
 - This means not more update timing with Java side.
 - Performance is greatyly improved.

In EmulatorActivity, changes onRestart/onStop to onResume/onPause.
Not quite sure why, but onStop would not be called anymore. But onPause
is. And the latter seems like the right lifecycle for this anyway.

Also had to remove the synchronization for taking the screenshot since
it's called from a different thread than run() and thus would otherwise
block.